+++ internal/provider/provider.go
@@ -243,6 +243,7 @@ func (p *OMVProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewSharedFolderResource,
 		NewRsyncJobResource,
+		NewWorkbenchSettingsResource,
 	}
 }